+++ Documents/ReleaseNotes.md
@@ -3,6 +3,7 @@
 ### [<- Go Back](../README.md)
 - `11.0.0-Preview-3.1`
    - `Map Controls`. You can now create/remove instances of basic atlas map controls via `AtlasInterop.Controls`.
+   - `IOptions<MapConfiguration>`. Added extension method `GetValue()`.
 - `11.0.0-Preview-3.0`
    - `Map Instances`. You can now create/remove instances of the atlas Map via `AtlasInterop.Factory`.
 - `11.0.0-Preview-2.0`